+++ library/MultiRegions/ExpList0D.cpp
@@ -70,8 +70,8 @@ namespace Nektar
             m_npoints = 1;
 
             // Set up m_coeffs, m_phys.
-            m_coeffs = Array<OneD, NekDouble> {m_ncoeffs, 0.0};
-            m_phys   = Array<OneD, NekDouble> {m_npoints, 0.0};
+            m_coeffs = Array<OneD, NekDouble> {size_t(m_ncoeffs), 0.0};
+            m_phys   = Array<OneD, NekDouble> {size_t(m_npoints), 0.0};
 
             LocalRegions::PointExpSharedPtr Point = MemoryManager<LocalRegions::PointExp>::AllocateSharedPtr(geom);
             (*m_exp).push_back(Point);
@@ -190,8 +190,8 @@ namespace Nektar
             // Set up m_coeffs, m_phys.
             if(DeclareCoeffPhysArrays)
             {
-                m_coeffs = Array<OneD, NekDouble> {m_ncoeffs,0.0};
-                m_phys   = Array<OneD, NekDouble> {m_npoints,0.0};
+                m_coeffs = Array<OneD, NekDouble> {size_t(m_ncoeffs),0.0};
+                m_phys   = Array<OneD, NekDouble> {size_t(m_npoints),0.0};
             }
         }
 
@@ -252,7 +252,7 @@ namespace Nektar
             Array<OneD, NekDouble>  &lengthsFwd,
             Array<OneD, NekDouble>  &lengthsBwd)
         {
-            int e_npoints;
+            size_t e_npoints;
 
             Array<OneD, NekDouble> locLeng;
             Array<OneD, NekDouble> lengintp;
