+++ library/MultiRegions/ExpList2D.cpp
@@ -222,8 +222,8 @@ namespace Nektar
             if (DeclareCoeffPhysArrays)
             {
                 // Set up m_coeffs, m_phys.
-                m_coeffs = Array<OneD, NekDouble> {m_ncoeffs, 0.0};
-                m_phys   = Array<OneD, NekDouble> {m_npoints, 0.0};
+                m_coeffs = Array<OneD, NekDouble> {size_t(m_ncoeffs), 0.0};
+                m_phys   = Array<OneD, NekDouble> {size_t(m_npoints), 0.0};
              }
 
             CreateCollections(ImpType);
@@ -331,8 +331,8 @@ namespace Nektar
             if (DeclareCoeffPhysArrays)
             {
                 // Set up m_coeffs, m_phys.
-                m_coeffs = Array<OneD, NekDouble> {m_ncoeffs, 0.0};
-                m_phys   = Array<OneD, NekDouble> {m_npoints, 0.0};
+                m_coeffs = Array<OneD, NekDouble> {size_t(m_ncoeffs), 0.0};
+                m_phys   = Array<OneD, NekDouble> {size_t(m_npoints), 0.0};
              }
 
             CreateCollections(ImpType);
@@ -754,8 +754,8 @@ namespace Nektar
             // Set up m_coeffs, m_phys.
             if(DeclareCoeffPhysArrays)
             {
-                m_coeffs = Array<OneD, NekDouble> {m_ncoeffs, 0.0};
-                m_phys   = Array<OneD, NekDouble> {m_npoints, 0.0};
+                m_coeffs = Array<OneD, NekDouble> {size_t(m_ncoeffs), 0.0};
+                m_phys   = Array<OneD, NekDouble> {size_t(m_npoints), 0.0};
             }
 
             CreateCollections(ImpType);
@@ -879,8 +879,8 @@ namespace Nektar
 
              // Set up m_coeffs, m_phys and offset arrays.
             SetCoeffPhysOffsets();
-            m_coeffs = Array<OneD, NekDouble> {m_ncoeffs, 0.0};
-            m_phys   = Array<OneD, NekDouble> {m_npoints, 0.0};
+            m_coeffs = Array<OneD, NekDouble> {size_t(m_ncoeffs), 0.0};
+            m_phys   = Array<OneD, NekDouble> {size_t(m_npoints), 0.0};
 
             CreateCollections(ImpType);
         }
