+++ library/MultiRegions/ExpList3D.cpp
@@ -76,8 +76,8 @@ namespace Nektar
             if (DeclareCoeffPhysArrays)
             {
                 // Set up m_coeffs, m_phys.
-                m_coeffs = Array<OneD, NekDouble> {m_ncoeffs, 0.0};
-                m_phys   = Array<OneD, NekDouble> {m_npoints, 0.0};
+                m_coeffs = Array<OneD, NekDouble> {size_t(m_ncoeffs), 0.0};
+                m_phys   = Array<OneD, NekDouble> {size_t(m_npoints), 0.0};
              }
 
             CreateCollections(ImpType);
@@ -171,8 +171,8 @@ namespace Nektar
             SetCoeffPhysOffsets();
 
             // Set up m_coeffs, m_phys.
-            m_coeffs = Array<OneD, NekDouble> {m_ncoeffs, 0.0};
-            m_phys   = Array<OneD, NekDouble> {m_npoints, 0.0};
+            m_coeffs = Array<OneD, NekDouble> {size_t(m_ncoeffs), 0.0};
+            m_phys   = Array<OneD, NekDouble> {size_t(m_npoints), 0.0};
 
             CreateCollections(ImpType);
         }
@@ -297,8 +297,8 @@ namespace Nektar
             SetCoeffPhysOffsets();
 
             // Set up m_coeffs, m_phys.
-            m_coeffs = Array<OneD, NekDouble> {m_ncoeffs, 0.0};
-            m_phys   = Array<OneD, NekDouble> {m_npoints, 0.0};
+            m_coeffs = Array<OneD, NekDouble> {size_t(m_ncoeffs), 0.0};
+            m_phys   = Array<OneD, NekDouble> {size_t(m_npoints), 0.0};
 
             CreateCollections(ImpType);
         }
@@ -425,8 +425,8 @@ namespace Nektar
             SetCoeffPhysOffsets();
 
             // Set up m_coeffs, m_phys.
-            m_coeffs = Array<OneD, NekDouble> {m_ncoeffs, 0.0};
-            m_phys   = Array<OneD, NekDouble> {m_npoints, 0.0};
+            m_coeffs = Array<OneD, NekDouble> {size_t(m_ncoeffs), 0.0};
+            m_phys   = Array<OneD, NekDouble> {size_t(m_npoints), 0.0};
 
             CreateCollections(ImpType);
         }
